src/test/java/net/ddellspe/day17/Day17Test.java
package net.ddellspe.day17;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

import org.junit.Test;

public class Day17Test {
  @Test
  public void providedInputTestPart1() {
    assertThat(Day17.part1("example.txt"), is(equalTo(45L)));
  }

  @Test
  public void solutionPart1() {
    System.out.println("Day 17 Part 1 Answer is: " + Day17.part1("input.txt"));
  }

  @Test
  public void providedInputTestPart2() {
    assertThat(Day17.part2("example.txt"), is(equalTo(112L)));
  }

  @Test
  public void solutionPart2() {
    System.out.println("Day 17 Part 2 Answer is: " + Day17.part2("input.txt"));
  }
}
